requirements.txt
requests == 2.32.3
langchain-core == 0.3.34
langchain-ollama == 0.2.3
langchainhub == 0.1.21
langchain == 0.3.18
langchain-community == 0.3.17
python-dotenv == 1.0.1
langchain-google-genai == 2.0.9
validators == 0.34.0
pydantic == 2.10.6
streamlit == 1.42.0